backend/requirements.txt
fastapi==0.135.1
uvicorn==0.38.0
sqlalchemy==2.0.43
psycopg2==2.9.10
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
aiofiles==25.1.0
python-dotenv==1.0.0
stripe==14.4.1
pydantic[email]
